% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BT_CV_Errors.R
\name{BT_cv_errors}
\alias{BT_cv_errors}
\title{Cross-validation errors.}
\usage{
BT_cv_errors(BT_cv_fit, cv.folds, folds)
}
\arguments{
\item{BT_cv_fit}{a \code{\link{BTCVFit}} object.}

\item{cv.folds}{a numeric corresponding to the number of folds.}

\item{folds}{a numerical vector containing the different \code{folds.id}. Note that if the latter was not defined by the user, those are randomly generated based on the \code{cv.folds} input.}
}
\value{
Vector containing the cross-validation errors w.r.t. the boosting iteration.
}
\description{
Function to compute the cross-validation error.
}
\details{
This function computes the the global cross-validation error as a function of the boosting iteration. Differently said, this measure is obtained by
computing the weighted average of out-of-fold errors.
}
\references{
D. Hainaut, J. Trufin and M. Denuit (2019). \dQuote{Effective Statistical Learning Methods for Actuaries, volume 1, 2 & 3,} \emph{Springer Actuarial}.
}
\seealso{
\code{\link{BT}}.
}
\author{
Gireg Willame \email{g.willame@detralytics.eu}

Julien Trufin \email{j.trufin@detralytics.eu}

Michel Denuit \email{m.denuit@detralytics.eu}
}
