% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BT_devTweedie.R
\name{BT_devTweedie}
\alias{BT_devTweedie}
\title{Deviance function for the Tweedie family.}
\usage{
BT_devTweedie(y, mu, tweedieVal, w = NULL)
}
\arguments{
\item{y}{a vector containing the observed values.}

\item{mu}{a vector contraining the fitted values.}

\item{tweedieVal}{a numeric representing the Tweedie Power. It has to be a positive numeric outside of the interval ]0,1[.}

\item{w}{a vector of weights. If no weights are needed, please}
}
\value{
A vector of individual deviance contribution.
}
\description{
Compute the deviance for the Tweedie family case.
}
\details{
This function computes the Tweedie related deviance. The latter is defined as
}
\references{
D. Hainaut, J. Trufin and M. Denuit (2019). \dQuote{Effective Statistical Learning Methods for Actuaries, volume 1, 2 & 3,} \emph{Springer Actuarial}.
}
\seealso{
\code{\link{BT}}, \code{\link{BT_call}}.
}
\author{
Gireg Willame \email{g.willame@detralytics.eu}

\emph{This package is inspired by the \code{gbm3} package. For more details, see \url{https://github.com/gbm-developers/gbm3/}}.
}
