% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BT_More.R
\name{BT_more}
\alias{BT_more}
\title{Perform additionnal boosting.}
\usage{
BT_more(BTFit_object, new.n.iter = 100, is.verbose = FALSE)
}
\arguments{
\item{BTFit_object}{a \code{\link{BTFit}} object.}

\item{new.n.iter}{number of new boosting iterations to perform.}

\item{is.verbose}{a logical specifying whether or not the additional fitting should run "noisely" with feedback on progress provided to the user.}
}
\value{
Returns a new \code{\link{BTFit}} object containing the initial call as well as the new iterations performed.
}
\description{
Method to perform additional iterations of the Boosting Tree algorithm, starting from an initial \code{\link{BTFit}} object.
This does not support further cross-validation. Moreover, this approach is only allowed if \code{keep.data=TRUE} in the original call.
}
\details{
Please note there is no seed set in this function. If user wants to obtain reproducible results, the upfront usage of \code{set.seed} is suggested.
}
\references{
D. Hainaut, J. Trufin and M. Denuit (2019). \dQuote{Effective Statistical Learning Methods for Actuaries, volume 1, 2 & 3}, \emph{Springer Actuarial}.
}
\seealso{
\code{\link{BT}}, \code{\link{BTFit}}.
}
\author{
Gireg Willame \email{g.willame@detralytics.eu}

\emph{This package is inspired by the \code{gbm3} package. For more details, see \url{https://github.com/gbm-developers/gbm3/}}.
}
