% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BT_Relative_Influence.R
\name{BT_relative_influence}
\alias{BT_relative_influence}
\title{Method for estimating the relative influence.}
\usage{
BT_relative_influence(BTFit_object, n.iter, rescale = FALSE, sort.it = FALSE)
}
\arguments{
\item{BTFit_object}{a \code{\link{BTFit}} object.}

\item{n.iter}{number of boosting iteration used for computations. If not provided, the function will perform a best guest approach to determine the optimal number of iterations. In fact,
if a test set was used during the fitting, the retained number of iterations is the one corresponding to the lowest test set error ; otherwise, if cross-validation was performed, the
number of iterations resulting in lowest cross-validation error will be used; otherwise, if the out-of-bag parameter was defined, the OOB error will be used to determine the optimal
number of iterations; otherwise, all iterations will be used.}

\item{rescale}{whether or not the results should be rescaled. Default set to \code{FALSE}.}

\item{sort.it}{whether or not the results should be (reverse) sorted. Default set to \code{FALSE}.}
}
\value{
Returns by default an unprocessed vector of estimated relative influences. If the \code{rescale} and \code{sort.it} arguments are used, it returns
a processed version of the same vector.
}
\description{
Helper function for computing the relative influence of each variable in the BT object.
}
\details{
This function is not intended for end-user use. It performs the relative influence computation and is called during the summary function.
Note that a permutation approach is not yet implemented.
}
\references{
D. Hainaut, J. Trufin and M. Denuit (2019). \dQuote{Effective Statistical Learning Methods for Actuaries, volume 1, 2 & 3,} \emph{Springer Actuarial}.
}
\seealso{
\code{\link{BT}}, \code{\link{BTFit}}.
}
\author{
Gireg Willame \email{g.willame@detralytics.eu}

Julien Trufin \email{j.trufin@detralytics.eu}

Michel Denuit \email{m.denuit@detralytics.eu}
}
