% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTCVFit.R
\name{BTCVFit}
\alias{BTCVFit}
\title{BTCVFit}
\value{
a list of \code{\link{BTFit}} object which each element corresponding to a specific BT fit on a particular fold
}
\description{
These are objects representing CV fitted boosting trees.
}
\details{
CV (Adaptive) Boosting Tree Model Object.
}
\section{Structure }{
 The following components must be included in a legitimate \code{BTCVFit} object.
}

\references{
D. Hainaut, J. Trufin and M. Denuit (2019). \dQuote{Effective Statistical Learning Methods for Actuaries, volume 1, 2 & 3,} \emph{Springer Actuarial}.
}
\seealso{
\code{\link{BT}}.
}
\author{
Gireg Willame \email{g.willame@detralytics.eu}

\emph{This package is inspired by the \code{gbm3} package. For more details, see \url{https://github.com/gbm-developers/gbm3/}}.
}
\keyword{methods}
