% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BT_CV_Predict.R
\name{predict.BTCVFit}
\alias{predict.BTCVFit}
\title{Predictions for CV fitted BT models.}
\usage{
\method{predict}{BTCVFit}(object, data, cv.folds, folds, best.iter.cv, ...)
}
\arguments{
\item{object}{a \code{\link{BTCVFit}} object containing CV BT models.}

\item{data}{the database on which one wants to predict the different CV BT models.}

\item{cv.folds}{a positive integer specifying the number of folds to be used in cross-validation of the BT fit.}

\item{folds}{vector of integers specifying which row of data belongs to which cv.folds.}

\item{best.iter.cv}{the optimal number of trees with a CV approach.}

\item{\dots}{not currently used.}
}
\value{
Returns a vector of predictions for each cv folds.
}
\description{
Compute predictions from cross-validated Boosting Trees model.
}
\details{
This function has not been coded for public usage but rather to assess the cross-validation performances.
}
\references{
D. Hainaut, J. Trufin and M. Denuit (2019). \dQuote{Effective Statistical Learning Methods for Actuaries, volume 1, 2 & 3,} \emph{Springer Actuarial}.
}
\seealso{
\code{\link{BT}}, \code{\link{BTFit}}.
}
\author{
Gireg Willame \email{g.willame@detralytics.eu}

\emph{This package is inspired by the \code{gbm3} package. For more details, see \url{https://github.com/gbm-developers/gbm3/}}.
}
