% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BT_Print.R
\name{print.BTFit}
\alias{print.BTFit}
\alias{print_iters_and_dist}
\alias{print_perf_measures}
\title{Printing functions.}
\usage{
\method{print}{BTFit}(x, ...)

print_iters_and_dist(x)

print_perf_measures(x)
}
\arguments{
\item{x}{a \code{\link{BTFit}} object.}

\item{\dots}{arguments passed to \code{print.default}.}
}
\description{
Functions to print the BT results.
}
\details{
Print the different input parameters as well as obtained results (best iteration/performance & relative influence) given the chosen approach.
}
\references{
D. Hainaut, J. Trufin and M. Denuit (2019). \dQuote{Effective Statistical Learning Methods for Actuaries, volume 1, 2 & 3}, \emph{Springer Actuarial}.
}
\seealso{
\code{\link{BT}}, \code{\link{BT_relative_influence}}
}
\author{
Gireg Willame \email{g.willame@detralytics.eu}

\emph{This package is inspired by the \code{gbm3} package. For more details, see \url{https://github.com/gbm-developers/gbm3/}}.
}
